+++ src/codec/mod.rs
@@ -7,8 +7,8 @@ mod wire;
 
 pub use wire::{
     BlockKindSkeleton, BlockSkeleton, BlockSkeletonInsert, ColumnAlignmentWire, DocOp, Envelope,
-    ListItemSkeleton, MAX_WIRE_NEST_DEPTH, OpBody, TextUnitWire, WIRE_VERSION,
-    insert_block_paragraph_is_empty,
+    ListItemSkeleton, MAX_WIRE_NEST_DEPTH, MovedTextUnitWire, OpBody, TextBlockKindWire,
+    TextUnitWire, WIRE_VERSION, insert_block_paragraph_is_empty,
 };
 
 use thiserror::Error;